+++ docs/CHANGELOG.md
@@ -6,12 +6,18 @@ All notable changes to DeterministicESPAsyncWebServer are documented here.
 
 ### Bug Fixes
 
+- fix(docs) README.md ([`e168e53`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/e168e5300c4a3c95ed7b726c766458509873c03d))
 - clamp delta-seconds with int64_t, not long (LLP64 overflow) ([`a33cefd`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/a33cefdee3f3c719ee19bc3b21f0911c7770794f))
 - stop test/coverage.xml going stale ([`f95a667`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/f95a6678d5ba05395352f1846e658610cac23654))
 - make the whole native matrix buildable off Linux ([`49fe992`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/49fe99241dd74b5aad238ae4476db4d094854423))
 
 ### CI / Build
 
+- update CHANGELOG.md [skip ci] ([`d1d3cae`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/d1d3cae0bc2f39e403998466fcb708f262dbf1a4))
+- update CHANGELOG.md [skip ci] ([`c635c37`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/c635c37f92f03f5239b05cc937fa71654a381de6))
+- update CHANGELOG.md [skip ci] ([`626b69f`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/626b69f8cd456c135e3176c42a29f913b668247c))
+- update test report + coverage [skip ci] ([`f72be48`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/f72be484f874a98a4bed38fcf14cf65d5a401cd4))
+- update CHANGELOG.md [skip ci] ([`58359fc`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/58359fcb73f0e6ecc7879afa2e3398e608bf07be))
 - update CHANGELOG.md [skip ci] ([`ee77749`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/ee77749615bded63aff58627c5c89f8a57d44c3e))
 - update CHANGELOG.md [skip ci] ([`5866f01`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/5866f01c3537c1a2b03c87a752d4fb1114b376e5))
 - gate the job on test results instead of discarding them ([`91adf56`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/91adf565af92ceaf141ac655069ad123d29a4728))
@@ -22,12 +28,18 @@ All notable changes to DeterministicESPAsyncWebServer are documented here.
 
 ### Changes
 
+- Merge branch 'main' of https://github.com/dstroy0/DeterministicESPAsyncWebServer ([`305e155`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/305e15564fc67e2aee2ac83775a3c4848d36e591))
+- Merge branch 'main' of https://github.com/dstroy0/DeterministicESPAsyncWebServer ([`504cf2f`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/504cf2f9c865f6736881eca2c069fa124598d1d0))
+- update(docs) README.md ([`9639f6d`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/9639f6d8fb3b467e8553f86a8d324d6484db9ee7))
 - Merge remote-tracking branch 'origin/main' ([`111aa05`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/111aa056758f4a195bacd7dae64b96644438343d))
 - Merge branch 'main' of https://github.com/dstroy0/DeterministicESPAsyncWebServer ([`0b67d9e`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/0b67d9ebbab199e0ef2b9fde1d5d06a73ad34613))
 - local branch-coverage tooling ([`c8b0414`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/c8b0414c36cf01597261761d059c763d638190de))
 
 ### Documentation
 
+- regenerate feature tables + configurator + build_opt.h + example index [skip ci] ([`37d5456`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/37d5456e62fcb58814a2b6b86854eeee1718d6bc))
+- update ESP32 build footprints [skip ci] ([`f596319`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/f596319783ab987e62ae4c378630c4c8de6c7da0))
+- regenerate feature tables + configurator + build_opt.h + example index [skip ci] ([`d9cac14`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/d9cac14e5ed2a37fee50aff1ad0c72113933e20c))
 - update ESP32 build footprints [skip ci] ([`fbbbad8`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/fbbbad862975b472e1ac34d135f869a6bf7ccc20))
 - regenerate feature tables + configurator + build_opt.h + example index [skip ci] ([`6bdfef0`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/6bdfef0ffd77a294e7b7611ab8e008785c928237))
 - update ESP32 build footprints [skip ci] ([`737ec03`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/737ec031002c44f415ffc4ac6ec8e2a3df5258c3))
@@ -37,6 +49,10 @@ All notable changes to DeterministicESPAsyncWebServer are documented here.
 
 ### Features
 
+- RCWL-0516 presence sensor + shared one-GPIO presence facade ([`882c38a`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/882c38ac782135418ab54875807083b309916e31))
+- support the HLK-LD2410B, and decode command ACKs ([`3ae0466`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/3ae0466be441afc6dede863038e433c45259f605))
+- FANUC Stream Motion (J519) UDP codec ([`bfbdcfd`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/bfbdcfdd6b70a7ed71123410aa2d1821feef4000))
+- feat(dpa/cpa) side channel analysis engine ([`7f5cf51`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/7f5cf51ac67b2368b5b93fa8ce39c76797889f3c))
 - EUROMAP 77 (OPC 40077) IMM_MES_Interface model ([`50c00a5`](https://github.com/dstroy0/DeterministicESPAsyncWebServer/commit/50c00a58b77f1adbcceee7c14b881c8a00f78784))
 
 ### Refactor